requirements.txt
Django==3.2
djangorestframework==3.12.4
django_rest_authtoken==2.1.4
pytest==6.2.4
pytest-django==4.4.0
Pillow==9.3.0
django-cors-headers==3.10.1
djangorestframework-camel-case==1.4.2
python-dotenv==1.1.0

--extra-index-url https://test.pypi.org/simple/
psycopg2-binary==2.9.3